move services package exclude from model
