+++ src/main/java/com/irdaislakhuafa/alterraacademyfinalproject/controllers/authors/AuthorAddressController.java
@@ -8,8 +8,8 @@ import javax.validation.Valid;
 import com.irdaislakhuafa.alterraacademyfinalproject.model.dtos.AddressDto;
 import com.irdaislakhuafa.alterraacademyfinalproject.model.requests.ApiChangeRequests;
 import com.irdaislakhuafa.alterraacademyfinalproject.model.requests.ApiTargetIdRequest;
-import com.irdaislakhuafa.alterraacademyfinalproject.model.services.AddressService;
-import com.irdaislakhuafa.alterraacademyfinalproject.model.services.AuthorService;
+import com.irdaislakhuafa.alterraacademyfinalproject.services.AddressService;
+import com.irdaislakhuafa.alterraacademyfinalproject.services.AuthorService;
 import com.irdaislakhuafa.alterraacademyfinalproject.utils.ApiMessage;
 import com.irdaislakhuafa.alterraacademyfinalproject.utils.ApiResponse;
 import com.irdaislakhuafa.alterraacademyfinalproject.utils.ApiValidation;
