+++ src/main/java/com/irdaislakhuafa/alterraacademyfinalproject/controllers/students/StudentController.java
@@ -1,0 +1,15 @@
+package com.irdaislakhuafa.alterraacademyfinalproject.controllers.students;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@RestController
+@RequestMapping(value = { "/students" })
+@RequiredArgsConstructor
+@Slf4j
+public class StudentController {
+
+}
